Correct display and redirect to another links
Hello.
By accepting this, the application will start to display links in which there is no text. At the moment, if you tweet only the link in the app, there will be an empty and non-working flash. Also, on rediirect, the webserver will send the client to the link from the flash.